+++ packages/openapi-parser/src/utils/openapi/utils/queueTask.ts
@@ -1,0 +1,11 @@
+import type { Queue, Task } from '../../../types'
+
+/**
+ * Add a new task to the existing queue
+ */
+export function queueTask<T extends Task[]>(queue: Queue, task: Task) {
+  return {
+    ...queue,
+    tasks: [...queue.tasks, task],
+  } as Queue<T>
+}
